+++ client/src/helpers/constants.js
@@ -357,7 +357,7 @@ export const NOT_FILTERED = 'NotFiltered';
 
 export const STATS_INTERVALS_DAYS = [0, 1, 7, 30, 90];
 
-export const QUERY_LOG_INTERVALS_DAYS = [1, 7, 30, 90];
+export const QUERY_LOG_INTERVALS_DAYS = [0.25, 1, 7, 30, 90];
 
 export const FILTERS_INTERVALS_HOURS = [0, 1, 12, 24, 72, 168];